requirements.txt
aiojobs==1.2.1
altair==4.2.2
anyio==4.4.0
astroid==2.12.13
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.2
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dill==0.3.6
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib-metadata==6.0.0
isort==5.11.4
Jinja2==3.1.2
jsonschema==4.17.3
lazy-object-proxy==1.9.0
lxml==4.9.2
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
numpy==1.24.1
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
platformdirs==2.6.2
protobuf==3.20.3
pyaes==1.6.1
pyarrow==11.0.0
pyasn1==0.6.0
pycodestyle==2.10.0
pydeck==0.8.0
Pygments==2.14.0
pylint==2.15.9
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-telegram-bot==21.3
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.3.1
rsa==4.9
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.3.2.post1
streamlit==1.17.0
Telethon==1.36.0
toml==0.10.2
tomlkit==0.11.6
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
wrapt==1.14.1
xmltodict==0.13.0
zipp==3.12.0
